SE2/SE3/SO3 - consolidate EPS, add dtype-conditioned EPS and add float32 unit tests (#220)
* Added mechanism to condition EPS on dtype for SE2 and SE3.

* Parameterized SE2 and SE3 unit tests so that dtype can be switched easily.

* Renamed EPS as NEAR_ZERO_EPS in SE2.

* update constants

* update test functions for lie groups

* test_se2 works

* set gt values to be double

* fixed the sign bug for SE3.log_map() at pi

* SE3.exp_map() and SE3.log_map() tests passed

* test_compose() passed for SE3

* test_transform_from_and_to works for SE3

* all the tests passed for SE3

* fixed a bug in test_compose

* update constants for Lie groups

* update test_se3.py

* refactored test_se3 and se3

* update test_so3 and so3

* fixed sign bug for SO3.log_map

* refactor _SO3_matrix_check

* add so2 matrix check

* add matrix check for SE2

* fixed a bug for quaternion check

* adjust constants for so3 check

+++ theseus/optimizer/tests/test_manifold_gaussian.py
@@ -72,7 +72,7 @@ def test_to():
     for i in range(10):
         mean, precision = random_manifold_gaussian_params()
         t = th.ManifoldGaussian(mean, precision=precision)
-        dtype = torch.float64 if np.random.random() < 0.5 else torch.long
+        dtype = torch.float64 if np.random.random() < 0.5 else torch.float32
         t.to(dtype)
 
         for var in t.mean:
